🧹 Refactor data handling in model classes to replace TorchDataBatch with NumpyDataBatch, improving compatibility with Numpy-based data processing.
